refactor layout and nav components
