+++ frontend/src/app/layout.tsx
@@ -1,9 +1,7 @@
-'use client'
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import { useContext } from 'react'
-import { useRouter } from 'next/navigation'
-import { AuthProvider, AuthContext } from './AuthContext'
+import { AuthProvider } from './AuthContext'
+import Nav from '@/components/Nav'
 import "./globals.css";
 
 const geistSans = Geist({
@@ -20,26 +18,6 @@ export const metadata: Metadata = {
   title: "Create Next App",
   description: "Generated by create next app",
 };
-
-function Nav() {
-  const { token, setToken } = useContext(AuthContext)
-  const router = useRouter()
-  const logout = () => {
-    setToken(null)
-    router.push('/')
-  }
-  return (
-    <nav style={{ padding: '1rem', display: 'flex', gap: '1rem', borderBottom: '1px solid #ccc' }}>
-      <a href="/">Home</a>
-      {token && <a href="/dashboard">Dashboard</a>}
-      {token && <a href="/connect">Connect</a>}
-      {token && <a href="/plans">Plans</a>}
-      {!token && <a href="/login">Login</a>}
-      {!token && <a href="/register">Register</a>}
-      {token && <button onClick={logout}>Logout</button>}
-    </nav>
-  )
-}
 
 export default function RootLayout({ children }: Readonly<{ children: React.ReactNode }>) {
   return (
